requirements.txt
ag-ui-protocol==0.1.8
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.58.2
anyio==4.9.0
argcomplete==3.6.2
attrs==25.3.0
blockbuster==1.5.25
boto3==1.39.9
botocore==1.39.9
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cohere==5.16.1
colorama==0.4.6
cryptography==44.0.3
distro==1.9.0
dnspython==2.7.0
eval_type_backport==0.2.2
executing==2.2.0
fastapi==0.116.1
fastavro==1.11.1
filelock==3.18.0
forbiddenfruit==0.1.4
frozenlist==1.7.0
fsspec==2025.7.0
google-auth==2.40.3
google-genai==1.26.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
griffe==1.7.3
groq==0.30.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jsonschema_rs==0.29.1
langchain==0.3.26
langchain-core==0.3.69
langchain-mongodb==0.6.2
langchain-text-splitters==0.3.8
langgraph==0.5.3
langgraph-api==0.2.100
langgraph-checkpoint==2.1.1
langgraph-checkpoint-mongodb==0.1.4
langgraph-cli==0.3.5
langgraph-prebuilt==0.5.2
langgraph-runtime-inmem==0.6.0
langgraph-sdk==0.1.74
langsmith==0.4.8
lark==1.2.2
logfire==3.25.0
logfire-api==3.25.0
markdown-it-py==3.0.0
mcp==1.12.0
mdurl==0.1.2
mistralai==1.9.2
motor==3.7.1
multidict==6.6.3
numpy==2.3.1
openai==1.97.0
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-instrumentation==0.56b0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.11.0
ormsgpack==1.10.0
packaging==25.0
powerline-status==2.7
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybreaker==1.4.0
pycparser==2.22
pydantic==2.11.7
pydantic-ai==0.4.4
pydantic-ai-slim==0.4.4
pydantic-evals==0.4.4
pydantic-graph==0.4.4
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.1
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.1.3
starlette==0.47.1
structlog==25.4.0
tavily-python==0.7.10
tenacity==8.5.0
tiktoken==0.9.0
tokenizers==0.21.2
tqdm==4.67.1
truststore==0.10.1
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
